requirements.txt
# Interface utilisateur web
gradio==4.44.0

# Requêtes HTTP pour l'API Groq
requests==2.31.0

# Connecteur MySQL
mysql-connector-python==8.2.0

# Manipulation de données
pandas==2.1.4

# Modules Python standards (déjà inclus avec Python)
# json - module standard
# re - module standard  
# os - module standard